+++ packages/component/src/Styles/StyleSet/SendBox.js
@@ -1,20 +1,26 @@
 export default function createSendBoxStyle({
   sendBoxBackground,
-  sendBoxHeight,
   sendBoxBorderBottom,
   sendBoxBorderLeft,
   sendBoxBorderRight,
-  sendBoxBorderTop
+  sendBoxBorderTop,
+  sendBoxHeight
 }) {
   return {
-    '& > .main': {
-      alignItems: 'stretch',
-      backgroundColor: sendBoxBackground,
-      borderBottom: sendBoxBorderBottom,
-      borderLeft: sendBoxBorderLeft,
-      borderRight: sendBoxBorderRight,
-      borderTop: sendBoxBorderTop,
-      minHeight: sendBoxHeight
+    '&.webchat__send-box': {
+      '& .webchat__send-box__button--align-bottom': { alignSelf: 'flex-end' },
+      '& .webchat__send-box__button--align-stretch': { alignSelf: 'stretch' },
+      '& .webchat__send-box__button--align-top': { alignSelf: 'flex-start' },
+
+      '& .webchat__send-box__main': {
+        alignItems: 'stretch',
+        backgroundColor: sendBoxBackground,
+        borderBottom: sendBoxBorderBottom,
+        borderLeft: sendBoxBorderLeft,
+        borderRight: sendBoxBorderRight,
+        borderTop: sendBoxBorderTop,
+        minHeight: sendBoxHeight
+      }
     }
   };
 }